llm-evaluation-system/automated_rt/requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
jinja2==3.1.6
starlette==0.46.1

# LLM Client Libraries
openai==1.69.0
python-dotenv==1.0.0
aiohttp==3.11.16

# Document Processing
pypdf==5.6.0
cryptography==45.0.1

# Data Handling
pydantic==2.11.1
numpy==2.2.4
scikit-learn==1.6.1

# Dependencies required by other packages
httpx==0.28.1  # Required by OpenAI client
typing_extensions==4.13.0  # Required by Pydantic and type hints